add the composer src/vendor folder
I don't see any good reason to leave this out of the repo.

Leaving it in the repo has benefits.  If something breaks with new
versions, or if composer can't pull in a dependency, there's a
working version in the previous commit.  Also, time is saved when
building because all the files don't have to be pulled in
separately.